Moving HttpUtils method to GvtUtils
