+++ dev/fattest.simplicity/bnd.bnd
@@ -115,5 +115,4 @@ fat.test.container.images: testcontainers/ryuk:0.6.0, testcontainers/sshd:1.1.0,
     org.seleniumhq.selenium:selenium-support;version=4.8.3,\
     org.bouncycastle:bcpg-jdk18on;version=1.75,\
     org.bouncycastle:bcpkix-jdk18on;version=1.75,\
-    org.bouncycastle:bcprov-jdk18on;version=1.75,\
-    com.ibm.ws.org.apache.httpcomponents;version=1.0.87
+    org.bouncycastle:bcprov-jdk18on;version=1.75
